refactor(structure): split into client/server, move assets, add routes
