+++ client/src/scripts/main.js
